+++ changelog.md
@@ -1,11 +1,33 @@
 # Changelog
 
-**./src/common/sub-interface/cmd-desc.js**
-* Added 'gridToImageDesc'
-	* Stored description for 'grid-to-image' command.
-	* Unlike other conversion commands, this is hard-coded.
-
----
-
-**./maze.js**
-* 'grid-to-image' command now has it's own description text.
+**./src/common/sub-interface/option-desc.js**
+* This file has been split into the new 'option-text' folder.
+	* op-output.js
+		* outputFolderDesc
+		* exportGraphDesc
+		* exportRawDesc
+		* replaceExistingFileDesc
+	* op-ignore-err.js
+		* ignoreTextParseDesc
+		* ignoreImageParseDesc
+		* writeIgnoreParseDesc
+	* op-img-config.js
+		* loadImageConfigDesc
+		* saveImageConfigDesc
+		* saveImageConvConfigDesc
+		* writeLoadSaveImageConfigDesc
+		* writeSaveImageConvConfigDesc
+	* op-img-parse.js
+		* tolerancePercentDesc
+		* tileSizeDesc
+		* imageOriginDescX
+		* imageOriginDescY
+		* writeOriginDesc
+	* op-img-colours.js
+		* wallColourDesc
+		* floorColourDesc
+		* writeColourDesc
+* This file now only contains:
+	* Header comment.
+	* Public exports.
+	* Placeholder text